+++ lib/runner.js
@@ -1,6 +1,5 @@
 var protractor = require('./protractor'),
     webdriver = require('selenium-webdriver'),
-    ConfigParser = require('./configParser'),
     path = require('path'),
     util = require('util'),
     q = require('q'),
@@ -51,191 +50,6 @@ util.inherits(Runner, EventEmitter);
 
 
 /**
- * Execute the Runner's test cases through Jasmine.
- *
- * @private
- * @param {Array} specs Array of Directory Path Strings
- * @param done A callback for when tests are finished.
- */
-Runner.prototype.runJasmine_ = function(specs, done) {
-  var minijn = require('minijasminenode'),
-      self = this;
-
-  require('../jasminewd');
-  /* global jasmine */
-
-  var RunnerReporter = function(emitter) {
-    this.emitter = emitter;
-  };
-
-  RunnerReporter.prototype.reportRunnerStarting = function() {};
-  RunnerReporter.prototype.reportRunnerResults = function() {};
-  RunnerReporter.prototype.reportSuiteResults = function() {};
-  RunnerReporter.prototype.reportSpecStarting = function() {};
-  RunnerReporter.prototype.reportSpecResults = function(spec) {
-    if (spec.results().passed()) {
-      this.emitter.emit('testPass');
-    } else {
-      this.emitter.emit('testFail');
-    }
-  };
-  RunnerReporter.prototype.log = function() {};
-
-  // On timeout, the flow should be reset. This will prevent webdriver tasks
-  // from overflowing into the next test and causing it to fail or timeout
-  // as well. This is done in the reporter instead of an afterEach block
-  // to ensure that it runs after any afterEach() blocks with webdriver tasks
-  // get to complete first.
-  jasmine.getEnv().addReporter(new RunnerReporter(this));
-
-  webdriver.promise.controlFlow().execute(function() {
-    self.runTestPreparers_();
-  }, 'run test preparers').then(function() {
-    var jasmineNodeOpts = self.config_.jasmineNodeOpts;
-    var originalOnComplete = self.config_.onComplete;
-    jasmineNodeOpts.onComplete = function(runner, log) {
-      if (originalOnComplete) {
-        originalOnComplete(runner, log);
-      }
-      done(runner.results());
-    };
-
-    minijn.addSpecs(specs);
-    minijn.executeSpecs(jasmineNodeOpts);
-  });
-};
-
-
-/**
- * Execute the Runner's test cases through Mocha.
- *
- * @private
- * @param {Array} specs Array of Directory Path Strings
- * @param done A callback for when tests are finished.
- */
-Runner.prototype.runMocha_ = function(specs, done) {
-
-  var Mocha = require('mocha'),
-      mocha = new Mocha(this.config_.mochaOpts),
-      self = this;
-
-  // Mocha doesn't set up the ui until the pre-require event, so
-  // wait until then to load mocha-webdriver adapters as well.
-  mocha.suite.on('pre-require', function() {
-    var mochaAdapters = require('selenium-webdriver/testing');
-    global.after = mochaAdapters.after;
-    global.afterEach = mochaAdapters.afterEach;
-    global.before = mochaAdapters.before;
-    global.beforeEach = mochaAdapters.beforeEach;
-
-    // The implementation of mocha's it.only uses global.it, so since that has
-    // already been wrapped we must avoid wrapping it a second time.
-    // See Mocha.prototype.loadFiles and bdd's context.it.only for more details.
-    var originalOnly = global.it.only;
-    global.it = mochaAdapters.it;
-    global.it.only = global.iit = originalOnly;
-
-    global.it.skip = global.xit = mochaAdapters.xit;
-  });
-
-  mocha.suite.on('pass', function() {
-    this.emit('testPass');
-  });
-
-  mocha.suite.on('fail', function() {
-    this.emit('testFail');
-  });
-
-  mocha.loadFiles();
-
-  webdriver.promise.controlFlow().execute(function() {
-    self.runTestPreparers_();
-  }, 'run test preparers').then(function() {
-
-    specs.forEach(function(file) {
-      mocha.addFile(file);
-    });
-
-    mocha.run(function(failures) {
-      if (self.config_.onComplete) {
-        self.config_.onComplete();
-      }
-      var resolvedObj = {
-        failedCount: failures
-      };
-
-      done(resolvedObj);
-    });
-  });
-};
-
-/**
- * Execute the Runner's test cases through Cucumber.
- *
- * @private
- * @param {Array} specs Array of Directory Path Strings
- * @param done A callback for when tests are finished.
- */
-Runner.prototype.runCucumber_ = function(specs, done) {
-  // TODO - add the event interface for cucumber.
-  var Cucumber = require('cucumber'),
-      self = this,
-      execOptions = ['node', 'node_modules/.bin/cucumber-js'],
-      cucumberResolvedRequire;
-
-  // Set up exec options for Cucumber
-  execOptions = execOptions.concat(specs);
-  if (self.config_.cucumberOpts) {
-
-    // Process Cucumber Require param
-    if (self.config_.cucumberOpts.require) {
-      cucumberResolvedRequire =
-          ConfigParser.resolveFilePatterns(self.config_.cucumberOpts.require);
-      if (cucumberResolvedRequire && cucumberResolvedRequire.length) {
-          execOptions = cucumberResolvedRequire.reduce(function(a, fn) {
-            return a.concat('-r', fn);
-          }, execOptions);
-      }
-    }
-
-    // Process Cucumber Tag param
-    if (Array.isArray(self.config_.cucumberOpts.tags)) {
-        for (var i in self.config_.cucumberOpts.tags) {
-            var tags = self.config_.cucumberOpts.tags[i];
-            execOptions.push('-t');
-            execOptions.push(tags);
-        }
-    } else if (self.config_.cucumberOpts.tags) {
-      execOptions.push('-t');
-      execOptions.push(self.config_.cucumberOpts.tags);
-    }
-
-    // Process Cucumber Format param
-    if (self.config_.cucumberOpts.format) {
-      execOptions.push('-f');
-      execOptions.push(self.config_.cucumberOpts.format);
-    }
-  }
-  global.cucumber = Cucumber.Cli(execOptions);
-
-  webdriver.promise.controlFlow().execute(function() {
-    self.runTestPreparers_();
-  }, 'run test preparers').then(function() {
-
-    global.cucumber.run(function(succeeded) {
-      if (self.config_.onComplete) {
-        self.config_.onComplete();
-      }
-      var resolvedObj = {
-        failedCount: succeeded ? 0 : 1
-      };
-      done(resolvedObj);
-    });
-  });
-};
-
-
-/**
  * Internal helper for abstraction of polymorphic filenameOrFn properties.
  * @private
  * @param {Array} source The Array that we'll be iterating through
@@ -271,9 +85,9 @@ Runner.prototype.registerTestPreparer = function(filenameOrFn) {
 
 /**
  * Executor of testPreparers
- * @private
+ * @public
  */
-Runner.prototype.runTestPreparers_ = function() {
+Runner.prototype.runTestPreparers = function() {
   this.runFilenamesOrFns_(this.preparers_);
 };
 
@@ -326,6 +140,15 @@ Runner.prototype.exit_ = function(exitCode) {
  */
 Runner.prototype.getConfig = function() {
   return this.config_;
+};
+
+
+/**
+ * Get the control flow used by this runner.
+ * @return {Object} WebDriver control flow.
+ */
+Runner.prototype.controlFlow = function() {
+  return webdriver.promise.controlFlow();
 };
 
 
@@ -392,25 +215,22 @@ Runner.prototype.run = function() {
 
     // Do the framework setup here so that jasmine and mocha globals are
     // available to the onPrepare function.
+    var frameworkPath = '';
     if (self.config_.framework === 'jasmine') {
-      self.runJasmine_.bind(self)(specs, function(result) {
-        deferred.resolve(result);
-      });
+      frameworkPath = './frameworks/jasmine.js';
     } else if (self.config_.framework === 'mocha') {
-      self.runMocha_.bind(self)(specs, function(result) {
-        deferred.resolve(result);
-      });
+      frameworkPath = './frameworks/mocha.js';
     } else if (self.config_.framework === 'cucumber') {
-      self.runCucumber_.bind(self)(specs, function(result) {
-        deferred.resolve(result);
-      });
-    } else if (self.config_.framework === 'simpleprint') {
-      console.log('Resolved spec files ' + util.inspect(specs));
-      deferred.resolve({failedCount: 0});
+      frameworkPath = './frameworks/cucumber.js';
+    } else if (self.config_.framework === 'debugprint') {
+      frameworkPath = './frameworks/debugprint.js';
     } else {
       throw new Error('config.framework (' + self.config_.framework +
           ') is not a valid framework.');
     }
+    require(frameworkPath).run(self, specs, function(result) {
+      deferred.resolve(result);
+    });
 
     return deferred.promise;
